requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
litellm==1.63.11
sqlalchemy==2.0.40
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
jinja2==3.1.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
